Makefile is now working fine
